+++ Tests/StreamChatTestTools/SpyPattern/Spy/ChannelListUpdater_Spy.swift
@@ -18,6 +18,9 @@ final class ChannelListUpdater_Spy: ChannelListUpdater, Spy {
     var resetChannelsQueryResult: Result<(synchedAndWatched: [ChatChannel], unwanted: Set<ChannelId>), Error>?
     
     @Atomic var markAllRead_completion: ((Error?) -> Void)?
+
+    @Atomic var startWatchingChannels_cids: [ChannelId] = []
+    @Atomic var startWatchingChannels_completion: ((Error?) -> Void)?
     
     func cleanUp() {
         update_queries.removeAll()
@@ -27,6 +30,9 @@ final class ChannelListUpdater_Spy: ChannelListUpdater, Spy {
         fetch_completion = nil
         
         markAllRead_completion = nil
+        
+        startWatchingChannels_cids.removeAll()
+        startWatchingChannels_completion = nil
     }
     
     override func update(
@@ -57,5 +63,10 @@ final class ChannelListUpdater_Spy: ChannelListUpdater, Spy {
     ) {
         record()
         resetChannelsQueryResult.map(completion)
+    }
+
+    override func startWatchingChannels(withIds ids: [ChannelId], completion: ((Error?) -> Void)?) {
+        startWatchingChannels_cids = ids
+        startWatchingChannels_completion = completion
     }
 }